Changed the location of "GetuserCategories" methods.
