+++ CRM-api/Controllers/User Module/UserCategoryController.cs
@@ -1,6 +1,8 @@
-﻿using CRM_api.DataAccess.Models;
+﻿using CRM_api.DataAccess.Helper;
+using CRM_api.DataAccess.Models;
 using CRM_api.Services.Dtos.AddDataDto.User_Module;
 using CRM_api.Services.IServices.User_Module;
+using CRM_api.Services.Services.User_Module;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CRM_api.Controllers.User_Module
@@ -16,6 +18,24 @@ namespace CRM_api.Controllers.User_Module
             _userCategoryService = userCategoryService;
         }
 
+        #region Get All User Category
+        [HttpGet("GetUserCatagories")]
+        public async Task<IActionResult> GetUserCatagories([FromQuery] string? search, [FromQuery] SortingParams? sortingParams)
+        {
+            try
+            {
+                var categories = await _userCategoryService.GetUserCategoriesAsync(search, sortingParams);
+                if (categories.Values.Count == 0)
+                    return BadRequest(new { Message = "Category not found." });
+
+                return Ok(categories);
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+        }
+        #endregion
 
         #region Add User Category
         [HttpPost("AddUserCategory")]
