+++ CRM_api.Services/IServices/User Module/IUserMasterService.cs
@@ -12,7 +12,6 @@ namespace CRM_api.Services.IServices.User_Module
         Task<ResponseDto<UserMasterDto>> GetUsersAsync(string filterString, string search, SortingParams sortingParams);
         Task<UserMasterDto> GetUserMasterByIdAsync(int id);
         Task<Dictionary<string, int>> GetUserCountAsync();
-        Task<ResponseDto<UserCategoryDto>> GetUserCategoriesAsync(string search, SortingParams sortingParams);
         Task<List<UserNameDto>> GetUsersByCategoryIdAsync(int categoryId, string search, SortingParams sortingParams);
         Task<TblUserCategoryMaster> GetCategoryByNameAsync(string name);
         int PanOrAadharExistAsync(int? id, string? pan, string? aadhar);
