+++ CRM_api.Services/Services/User Module/UserMasterService.cs
@@ -62,21 +62,6 @@ namespace CRM_api.Services.Services.User_Module
         }
         #endregion
 
-        #region Get All User Category
-        public async Task<ResponseDto<UserCategoryDto>> GetUserCategoriesAsync(string search, SortingParams sortingParams)
-        {
-            var catagories = await _userMasterRepository.GetUserCategories(search, sortingParams);
-            var mapCatagories = _mapper.Map<ResponseDto<UserCategoryDto>>(catagories);
-
-            foreach (var catagory in catagories.Values)
-            {
-                catagory.CatName = catagory.CatName.ToLower();
-            }
-
-            return mapCatagories;
-        }
-        #endregion
-
         #region Get All User By Category Id
         public async Task<List<UserNameDto>> GetUsersByCategoryIdAsync(int categoryId, string search, SortingParams sortingParams)
         {
